Copy jlibtool from upstream repository
Many projects use libtool, which is a cross-platform build tool
written in a shell script.  It is incredibly slow, however, so
replacing it with an executable may be a good idea.  The "jlibtool"
program from the Apache project has most of the functionality
of libtool, is nearly as portable, and is much faster.

Taken from http://svn.apache.org/repos/asf/apr/apr/trunk/build/jlibtool.c