Add ".libs/" in the correct place.
The earlier code added ".libs/" to the *start* of a filename.  If the
filename contained a "/", this resulted in "foo/bar.lo -> .libs/foo/bar.lo",
which was wrong.  It should instead map "foo/bar.lo -> foo/.libs/bar.lo"

This is especially important when the path contains "../".

There may be cases (e.g. cygwin) where the .libs directory isn't added
correctly.  But they should now be easy to fix.