Autodetect --mode
via magic binaries CC and LINK, LINK.c, and LINK.cxx

If we find them, we replace them with the correct CC, LINK, etc.
which were defined when jlibtool was compiled.

This means you need to learn FEWER command-line options to use
jlibtool.  Instead of doing "jlibtool --mode=compile $(CC)",
you just do "jlibtool CC ..."  Or, re-define the CC macro
so that "CC=CC".

Note that it's also possible to wrap this via other macros:

	CC=jlibtool --module=compile gcc ....

But this new method seems to be much easier to understand