Set default install path for libraries.
via the LIBDIR macro.  If the user asks to build a shared
library, he doesn't need to set -rpath.  Instead, we set it
to a default value.

That behavior is much friendlier than exiting with an error
when the user does "-shared" but not "-rpath ...".  It's also MUCH
better than the original behavior.  That silently ignored -shared
if there was no -rpath, and instead built a static library.  Ugh.