Create real output library.
If the output library is NOT a ".la", then the previous code
would still write the libtool path information to it.  The result
would be that the file was created with the wrong contents.

This changes makes it look at the filename extension.  If it's NOT
a ".la", then the output is created as a symlink to the real file
in the ".libs" directory.