Set --mode=execute automatically if we're symlinked
We have a directory:

jlibtool
program -> jlibtool
.libs/program

When we run "progam", jlibtool notices that it's NOT being called
as "jlibtool", and that there's a ".libs/program".  In that case,
it assumes you really wanted "jlibtool --mode=execute program..."
and it runs the program.

This is much easier than the libtool shell script wrapper

It means that you can't really do "gdb --args ./program ..."
as GDB will attach to jlibtool, and not to .libs/program.

But that doesn't work for libtool, either.  Instead you
have to do "libtool --mode=execute gdb ...".  That extended
syntax still works with jlibtool